Translation on same location works like a charm!

+++ TheExpanse/ExpanseHelper.h
@@ -7,9 +7,9 @@ class ExpanseHelper
 {
 public:
     void printMatrix(const Matrix& m);
-    Matrix getTranslationMatrix(int xChange, int yChange);
-    Matrix getScalingMatrix(int xScale, int yScale);
-    TwoDObject scaleOnLocation(TwoDObject& object, int xScale, int yScale);
+    Matrix getTranslationMatrix(double xChange, double yChange);
+    Matrix getScalingMatrix(double xScale, double yScale);
+    TwoDObject scaleOnLocation(TwoDObject& object, double xScale, double yScale);
 
 private:
